+++ src/main/java/com/packsendme/lib/tolls/response/dto/TollsCountryResponse_Dto.java
@@ -17,12 +17,11 @@ public class TollsCountryResponse_Dto implements Serializable {
 	public String name_country;
 	public int toll_amount;
 	public Double average_price_toll;
-	public Double distance_country;
+	public int distance_country;
 	public String currency_price;
 	
-	
 	public TollsCountryResponse_Dto(String name_country, int toll_amount, Double average_price_toll,
-			Double distance_country, String currency_price) {
+			int distance_country, String currency_price) {
 		super();
 		this.name_country = name_country;
 		this.toll_amount = toll_amount;
@@ -31,6 +30,7 @@ public class TollsCountryResponse_Dto implements Serializable {
 		this.currency_price = currency_price;
 	}
 	
+
